Move SQL CVEs to their own check
